Eliminate intermediate data types and callbacks
Since tm_ctags is inside tag manager, we have direct access to TMTag so
we can eliminate the auxiliary ctagsTag type. Similarly, we can access
tag manager structures directly from ctags writer callbacks so we can
eliminate tagEntryFunction and rescanFailedFunction too. The rest of
the patch just moves code around so we do the things directly at the
right place.
