+++ src/tagmanager/tm_ctags.c
@@ -22,6 +22,8 @@
 #include "xtag.h"
 #include "entry_p.h"
 
+#include "tm_tag.h"
+
 #include <stdio.h>
 #include <string.h>
 #include <errno.h>
@@ -44,9 +46,7 @@ tagWriter geanyWriter = {
 	.type = WRITER_U_CTAGS /* not really but we must use some of the builtin types */
 };
 
-static tagEntryFunction geanyTagEntryFunction = NULL;
-static rescanFailedFunction geanyRescanFailedFunction = NULL;
-static void *geanyUserData = NULL;
+static TMSourceFile *current_source_file = NULL;
 
 
 static bool nofatalErrorPrinter (const errorSelection selection,
@@ -85,38 +85,127 @@ static void enableAllLangKinds()
 }
 
 
-static void initCtagsTag(ctagsTag *tag, const tagEntryInfo *info)
+/*
+ Initializes a TMTag structure with information from a ctagsTag struct
+ used by the ctags parsers. Note that the TMTag structure must be malloc()ed
+ before calling this function.
+ @param tag The TMTag structure to initialize
+ @param file Pointer to a TMSourceFile struct (it is assigned to the file member)
+ @param tag_entry Tag information gathered by the ctags parser
+ @return TRUE on success, FALSE on failure
+*/
+static gboolean init_tag(TMTag *tag, TMSourceFile *file, const tagEntryInfo *tag_entry)
 {
-	tag->name = info->name;
-	tag->signature = info->extensionFields.signature;
-	tag->scopeName = info->extensionFields.scopeName;
-	tag->inheritance = info->extensionFields.inheritance;
-	tag->varType = info->extensionFields.typeRef[1];
-	tag->access = info->extensionFields.access;
-	tag->implementation = info->extensionFields.implementation;
-	tag->kindLetter = getLanguageKind(info->langType, info->kindIndex)->letter;
-	tag->isFileScope = info->isFileScope;
-	tag->lineNumber = info->lineNumber;
-	tag->lang = GEANY_LANG(info->langType);
+	TMTagType type;
+	guchar kind_letter;
+	TMParserType lang;
+
+	if (!tag_entry)
+		return FALSE;
+
+	lang = GEANY_LANG(tag_entry->langType);
+	kind_letter = getLanguageKind(tag_entry->langType, tag_entry->kindIndex)->letter;
+	type = tm_parser_get_tag_type(kind_letter, lang);
+	if (file->lang != lang)  /* this is a tag from a subparser */
+	{
+		/* check for possible re-definition of subparser type */
+		type = tm_parser_get_subparser_type(file->lang, lang, type);
+	}
+
+	if (!tag_entry->name || type == tm_tag_undef_t)
+		return FALSE;
+
+	tag->name = g_strdup(tag_entry->name);
+	tag->type = type;
+	tag->local = tag_entry->isFileScope;
+	tag->pointerOrder = 0;	/* backward compatibility (use var_type instead) */
+	tag->line = tag_entry->lineNumber;
+	if (NULL != tag_entry->extensionFields.signature)
+		tag->arglist = g_strdup(tag_entry->extensionFields.signature);
+	if ((NULL != tag_entry->extensionFields.scopeName) &&
+		(0 != tag_entry->extensionFields.scopeName[0]))
+		tag->scope = g_strdup(tag_entry->extensionFields.scopeName);
+	if (tag_entry->extensionFields.inheritance != NULL)
+		tag->inheritance = g_strdup(tag_entry->extensionFields.inheritance);
+	if (tag_entry->extensionFields.typeRef[1] != NULL)
+		tag->var_type = g_strdup(tag_entry->extensionFields.typeRef[1]);
+	if (tag_entry->extensionFields.access != NULL)
+		tag->access = tm_source_file_get_tag_access(tag_entry->extensionFields.access);
+	if (tag_entry->extensionFields.implementation != NULL)
+		tag->impl = tm_source_file_get_tag_impl(tag_entry->extensionFields.implementation);
+	if ((tm_tag_macro_t == tag->type) && (NULL != tag->arglist))
+		tag->type = tm_tag_macro_with_arg_t;
+	tag->file = file;
+	/* redefine lang also for subparsers because the rest of Geany assumes that
+	 * tags from a single file are from a single language */
+	tag->lang = file->lang;
+	return TRUE;
+}
+
+
+/* add argument list of __init__() Python methods to the class tag */
+static void update_python_arglist(const TMTag *tag, TMSourceFile *current_source_file)
+{
+	guint i;
+	const char *parent_tag_name;
+
+	if (tag->type != tm_tag_method_t || tag->scope == NULL ||
+		g_strcmp0(tag->name, "__init__") != 0)
+		return;
+
+	parent_tag_name = strrchr(tag->scope, '.');
+	if (parent_tag_name)
+		parent_tag_name++;
+	else
+		parent_tag_name = tag->scope;
+
+	/* going in reverse order because the tag was added recently */
+	for (i = current_source_file->tags_array->len; i > 0; i--)
+	{
+		TMTag *prev_tag = (TMTag *) current_source_file->tags_array->pdata[i - 1];
+		if (g_strcmp0(prev_tag->name, parent_tag_name) == 0)
+		{
+			g_free(prev_tag->arglist);
+			prev_tag->arglist = g_strdup(tag->arglist);
+			break;
+		}
+	}
 }
 
 
 static int writeEntry (tagWriter *writer, MIO * mio, const tagEntryInfo *const tag)
 {
-	ctagsTag t;
+	TMTag *tm_tag = tm_tag_new();
 
 	getTagScopeInformation((tagEntryInfo *)tag, NULL, NULL);
-	initCtagsTag(&t, tag);
-	geanyTagEntryFunction(&t, geanyUserData);
+
+	if (!init_tag(tm_tag, current_source_file, tag))
+	{
+		tm_tag_unref(tm_tag);
+		return 0;
+	}
+
+	if (tm_tag->lang == TM_PARSER_PYTHON)
+		update_python_arglist(tm_tag, current_source_file);
+
+	g_ptr_array_add(current_source_file->tags_array, tm_tag);
 
 	/* output length - we don't write anything to the MIO */
 	return 0;
 }
 
 
-static void rescanFailed (tagWriter *writer, unsigned long validTagNum)
+static void rescanFailed (tagWriter *writer, unsigned long valid_tag_num)
 {
-	geanyRescanFailedFunction (validTagNum, geanyUserData);
+	GPtrArray *tags_array = current_source_file->tags_array;
+
+	if (tags_array->len > valid_tag_num)
+	{
+		guint i;
+		for (i = valid_tag_num; i < tags_array->len; i++)
+			tm_tag_unref(tags_array->pdata[i]);
+		g_ptr_array_set_size(tags_array, valid_tag_num);
+	}
 }
 
 
@@ -147,9 +236,7 @@ extern void ctagsInit(void)
 
 
 extern void ctagsParse(unsigned char *buffer, size_t bufferSize,
-	const char *fileName, const int language,
-	tagEntryFunction tagCallback, rescanFailedFunction rescanCallback,
-	void *userData)
+	const char *fileName, const int language, TMSourceFile *source_file)
 {
 	if (buffer == NULL && fileName == NULL)
 	{
@@ -157,9 +244,7 @@ extern void ctagsParse(unsigned char *buffer, size_t bufferSize,
 		return;
 	}
 
-	geanyTagEntryFunction = tagCallback;
-	geanyRescanFailedFunction = rescanCallback;
-	geanyUserData = userData;
+	current_source_file = source_file;
 
 	geanyCreateTags(buffer, bufferSize, fileName, CTAGS_LANG(language));
 }
