+++ src/tagmanager/tm_ctags.h
@@ -12,33 +12,11 @@
 #include <stdlib.h>
 #include <stdbool.h>
 
-typedef struct {
-	const char *name;
-	const char *signature;
-	const char *scopeName;
-	const char *inheritance;
-	const char *varType;
-	const char *access;
-	const char *implementation;
-	char kindLetter;
-	bool isFileScope;
-	unsigned long lineNumber;
-	int lang;
-} ctagsTag;
-
-/* Callback invoked for every tag found by the parser. The return value is
- * currently unused. */
-typedef bool (*tagEntryFunction) (const ctagsTag *const tag, void *userData);
-
-/* Callback invoked when rescan fails telling us how many of the passed tags are valid */
-typedef void (*rescanFailedFunction) (unsigned long validTagNum, void *userData);
-
+#include "tm_source_file.h"
 
 extern void ctagsInit(void);
 extern void ctagsParse(unsigned char *buffer, size_t bufferSize,
-	const char *fileName, const int language,
-	tagEntryFunction tagCallback, rescanFailedFunction rescanCallback,
-	void *userData);
+	const char *fileName, const int language, TMSourceFile *source_file);
 extern const char *ctagsGetLangName(int lang);
 extern int ctagsGetNamedLang(const char *name);
 extern const char *ctagsGetLangKinds(int lang);
